Fix front-spa not building
