+++ front/lib/reinforcement/run_reinforced_analysis.ts
@@ -5,11 +5,9 @@ import type { LLM } from "@app/lib/api/llm/llm";
 import type { LLMEvent } from "@app/lib/api/llm/types/events";
 import type { LLMStreamParameters } from "@app/lib/api/llm/types/options";
 import { getLlmCredentials } from "@app/lib/api/provider_credentials";
-import {
-  getLargeWhitelistedModel,
-  getLargeWhitelistedModelWithBatchMode,
-} from "@app/lib/assistant";
+import { getLargeWhitelistedModel } from "@app/lib/assistant";
 import type { Authenticator } from "@app/lib/auth";
+import { getLargeWhitelistedModelWithBatchMode } from "@app/lib/reinforcement/models";
 import {
   hasSuggestionSelfConflict,
   pruneConflictingSkillEditSuggestions,
